allow for skip values larger than 2^32

+++ virvo/tools/vconv/vvconv.cpp
@@ -1823,7 +1823,7 @@ bool vvConv::parseCommandLine(int argc, char** argv)
         cerr << "Skip value missing." << endl;
         return false;
       }
-      rawSkip = atoi(argv[arg]);
+      rawSkip = atoll(argv[arg]);
     }
 
     else if (vvToolshed::strCompare(argv[arg], "-loadcpt")==0)
